requirements.txt
line-bot-sdk
python-dotenv
fastapi
uvicorn
openai
python-multipart
aiohttp
pydantic
PyYAML
Pillow
